+++ src/components/auth/register/Register.tsx
@@ -9,9 +9,9 @@ import Button from '@/components/common/Button';
 import Flex from '@/components/common/Flex';
 import Input from '@/components/common/Input';
 import Logo from '@/components/common/Logo';
-import { useModalState } from '@/components/common/Modal/useModalState';
 import { tilyLinks } from '@/constants/links';
 import { NAME_REGEX, PASSWORD_REGEX } from '@/constants/regex';
+import { useModalState } from '@/hooks/common/useModalState';
 import RegisterModal from './RegisterModal';
 
 interface RegisterFormInput {
