+++ src/components/common/Modal/index.stories.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import type { Meta } from '@storybook/react';
 import Button from '@/components/common/Button';
 import Modal from '@/components/common/Modal';
-import { useModalState } from '@/components/common/Modal/useModalState';
+import { useModalState } from '@/hooks/common/useModalState';
 
 export default {
   component: Modal,
